src: deploy: Fix wrong progress bar
Some config files that do not enable CONFIG_MODULE_SIG do not have the
extra SIGN output, making the progress bar display only half of the
progress. This commit fixes this issue by first checking the
CONFIG_MODULE_SIG=y option in the config; If it is not enabled, it drops
the 2 factor multiply.

Tested-by: Melissa Wen <melissa.srw@gmail.com>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>