src: build: Add olddefconfig to build
We tried to add olddefconfig as a step to be executed before building
the kernel to avoid the interactive menu. However, the first
implementation did not work well in a cross-compiled environment, and we
had to revert it. This commit brings it back and reworks the build tests
to make them more readable.

Closes #393

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>