src: deploy: Add support for Rasbian 64 bits
For deploying a new kernel to a Raspberry Pi with 32 bits, we need to
copy the dtb files by using:

 cp arch/arm/boot/dts/*.dtb <PATH>

In the 64 bits scenario we have to use:

 cp arch/arm64/boot/dts/broadcom/*.dtb <PATH>

This tiny difference raised the attention that we need to handle dtb
files deploy a little bit more carefully since any vendor can specify
any boot layout (at least, this is what I understood). For this reason,
this commit introduces a new config option named dtb_copy_pattern, which
gives the developers the flexibility to tell kw how it should copy these
dtbs files.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>