setup: Change the order of the installation steps
When installing kw without pip installed in the system, the following
error occurs:

./setup.sh: line 78: pip: command not found

That happens because setup.sh runs a pip command before checking and
installing the package dependencies -- pip included.

This commit simply changes the order of the steps, so that pip is not
used before it is installed.

Signed-off-by: Magali Lemes <magalilemes00@gmail.com>