src: deploy: Use kernel_img_name for the kernel image
In the kworkflow.config file, we have a variable named kernel_img_name
used to define the binary kernel prefix used during the deploy. Even
though we have this variable, we were ignoring it. As a result, when we
deploy a new kernel image to a rasp system, we see a kernel image with a
vmlinuz prefix. This commit fixes this issue by stopping to ignore that
config variable and only set vmlinuz if that variable is empty.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>