src: kernel_config_manager: Integrate with kw_db
kw kernel-config-manager uses metadata files to manage the database
management of kernel configuration files.

This commit changes this by integrating kernel-config-manager with kw_db
(that uses sqlite underneath). This is a more scalable approach and
easier to maintain as some logic of data management is shifted to kw_db.
The actual .config file is still stored in the user filesystem and just
the metadata files were transferred to the database.

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>