src: kwlib: check if the context is inside a git worktree
In the functions get_git_config_regex and get_all_git_config, git produce an
error when kw mail -l is running outside a git worktree.

In order to avoid such error, the git config function checks if the context
is inside a git worktree and changes the scope to global.

As this commit changes the behavior of the functions, the tests were
updated to cover the case where the user is inside a git tree and
outside it.

Closes #551

Signed-off-by: Maíra Canal <maira.canal@usp.br>