github: workflow: Fix CI failure caused by sphinx-book-theme
A sphinx-book-theme maintainer changed the way of getting the theme path.

Ref: https://github.com/executablebooks/sphinx-book-theme/blame/940d562fad0c8e9d4056b575c68d972ac0f9b2f4/src/sphinx_book_theme/__init__.py#L26

For jinja2>=3, it changes the way the variable is wrapped.

Ref: https://github.com/pallets/jinja/blob/5962edeb271d93687eb93f32d53ffe53f86871e0/src/jinja2/loaders.py#L184

This commit fixes the CI failure by updating jinja2 to the version that can
work with sphinx-book-theme in `pip.dependencies`.

The way to detect if a package is installed has been updated to use the
tools provided by python

A pair of double quotes encloses the package name to prevent the shell from
misrecognizing angle brackets in `setup.sh`.

Closes #583

Signed-off-by: Haiqin Cui <i@18kas.com>