src: deploy: Use INSTALL_MOD_STRIP by default
Usually, I work with the same set of test systems and kernel branches.
However, I had to switch hardware and branch recently, but I kept the
same config file since it was the same hardware family. Usually, I have
the CONFIG_DEBUG_INFO option set to yes, and I don't have any issue with
that; but for some reason, it started to generate a huge initramfs
(>256Mb) which caused boot failure. After a lot of research, I found
this page:

https://01.org/linuxgraphics/gfx-docs/drm/kbuild/kbuild.html

I also realize that the Debian package uses this option by default. For
this reason, I decided to add this as a default in the kw, and I also
added a warning in case the user has this option and explicitly disabled
the strip from the kw config file.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>