src: deploy: Fix sudo issue in local deploy
Some users reported that `kw bd --local` and `kw d --local` had some
weird behavior during the deploy. Something that looks like this:

    * Preparing modules
    [sudo] password for siqueira: [>

    [>

As a result, sometimes, users type the wrong password and have to type
it multiple times. This problem happens due to the pipe between the
modules_install command and the pv command. This commit fixes this issue
by running a simple sudo command before the task that requires the pipe
utilization.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>