src: plugins: kernel_install: utils: Enable write to a non-writable fs
SteamOS and other OSes protect the rootfs. If we want to enable kw
deploy for these types of OSes, we need to add support for making the OS
accept new kernels. This commit adds the required code to check if a
partition is writable or not and takes the necessary action to add write
permission.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>