src: report: separate report output per date
Previously when creating a report that spanned multiple days, if you had
entries of the same tag on multiple of those days, the output would put
them together and it would be hard to now to which date each entry
belonged. Now every new date adds a date line above the entries, making
for simplified viewing.

The previous implementation would output a report like this:
```
 - Total focus time: 01:30:00
 - Total repetitions: 5

Summary:
 * [13:53:53-14:23:53][30m]: Some description text
 * [17:47:42-18:17:42][30m]
 * [13:08:04-13:23:04][15m]: Some description text 2
 * [19:38:23-19:43:23][5m]
 * [20:30:41-20:40:41][10m]
```

The new implementation will produces the following report with the same
data:
```
 - Total focus time: 01:30:00
 - Total repetitions: 5

Summary:
 - 2021/09/08
   * [13:53:53-14:23:53][30m]: Some description text
   * [17:47:42-18:17:42][30m]
 - 2021/09/09
   * [13:08:04-13:23:04][15m]: Some description text 2
 - 2021/09/11
   * [19:38:23-19:43:23][5m]
   * [20:30:41-20:40:41][10m]
```

Closes #626

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@alumni.usp.br>