src: deploy: fix errors in listing kernels
The command to list installed kernels does some checks and setups that
are meant to be done when deploying a new kernel, but are unnecessary
when the intention is a different one, which ends up throwing some
errors. This commit ensures that the command to list kernels is run
before the deploy-related set-up, and also adds a missing argument from
the 'list_installed_kernel' function.

Signed-off-by: Magali Lemes <magalilemes00@gmail.com>