src: report: fix descriptions with commas
Since the data fields in the pomodoro files are separated by commas,
when `cut -d ',' -f4` was used to retrieve the descriptions, if they
contained one or more commas, any part of the description that was after
the first comma would be left out. Now the description is retrieved as
the complement of the rest of the data.

Closes #495

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@alumni.usp.br>