src: config: Introduce config option
Whenever users want to change a permanent config option, they have to
open the config file, which is not friendly behavior. This commit
introduces a new option named config, which provides a CLI for dealing
with our config file. Notice that I tried to make it as close as
possible to the git config option.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>