setup: add global config file
Some of kw's functions don't need to be used inside of a kernel tree,
for example, the `kw pomodoro` command that can be run on any folder,
but depend on values present in the config files. This means that we
need a global config file to ensure these functions work correctly.

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@alumni.usp.br>