src: configm: Rename configm to kernel-config-manager
We plan to add a config interface similar to the git config option to
make it easier for new users to learn how to set up kw. However, we are
using the name configm for the kernel config manager feature, which will
be super confusing for users. This commit address this issue by renaming
configm to kernel-config-manager.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>