src: kw_config_loader: Rework load configuration
We are moving toward a direction where we have fine control over which
config option we want to load per feature. For example, we don't need to
load codestyle or maintainers options in the build feature. This commit
reworks part of the config to make other options only load what they
need.

Closes #641

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>