src: explore: Add parameter amount check
Without checking the number of arguments can cause unexpected arguments
to be passed into the function.

The `explore_git_log` function in the `explore.sh` file will treat the
third parameter `$3` (extra parameters given by the user) as a `flag`
parameter and pass it to `cmd_manager`.

Before the commit f16c80, this would cause an error as said
in issue #409.

After the commit f16c80, the error will not occur, but the flag was
still passed with a wrong value.

This commit will output an error when the user gives the wrong amount of
parameters.

Closes #409

Signed-off-by: Haiqin Cui <i@18kas.com>