src: report: fix per tag hours of focus
The hours of focus per tag were previously displayed the output of using
the `date` command to convert the value from seconds, this had the side
effect of being capped at 24h. This is because the `date` command
converts seconds as the amount of seconds since the epoch, and so
returns the time of the corresponding date.

Closes #598

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@alumni.usp.br>