src: Adapt multiple part of kw to use new remote
I'm adapting kw to use the ssh config file as a default remote
manipulation, and this commit is one step in this direction since it
updates multiple features to use that. To keep this commit under
control, this change only ensures that all tests pass; however, the next
sequence of commits will laser focus on each feature that relies on ssh.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>