src: build: hide error generated by make
When CONFIG_X86_X32 is enabled but the corresponding tool is not
installed, an error `CONFIG_X86_X32 enabled but no binutils support`
will be reported when executing `make kernelrelease`, although no error
was reported when compiling the kernel.

Since this build_info function only gets the current kernel version by
executing the `make kernelrelease` command, and this error does not
affect the getting process of the version, this commit simply hides the
error message.

Closes #398

Signed-off-by: Haiqin Cui <i@18kas.com>