src: mail: create recipient groups
This feature enables the user to set a file containing a group of email
addresses to send the emails. The file can be created in the `etc`
directory if they are meant to be global, or in the `.kw` folder if they
are meant to be local to the current repository, in both cases thefile
should be under a `mail_groups` folder. In the future these files will
be handled by kw.

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@alumni.usp.br>