src: env: Introduce env files
When working with the Linux kernel, having multiple setups in the same
repository is common. For example, a developer that works with 64-bit
builds might want to check 32-bit compilation. To alleviate this type of
overhead, we introduce a new feature named 'kw env' to manage different
setups under the same repository. In particular, this commit only add
the create and list options, the next set of commits will add more
features and refine this first change.

Related to #114

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>