src: report: fix path validation in the report parameter.
When generating a report file, an error occurs if the user
specifies only the directory, the error occurs because a standard
output file is not created.

How to reproduce the error:

Execute `kw report -o ./`

This commit implements validation for the given path and auto-generate
a file name if a directory is given. After this commit, the user can
create an output file containing the Pomodoro and statistics data
in any desired directory.

Closes: #532

Signed-off-by: Aquila Macedo <aquilamacedo@riseup.net>