src: kernel_config_manager: replace undefined function call
When using 'kw --save | -s' to add a new .config file to kw management,
if the 'git commit' command fails, the undefined function 'fail()' is
called, generating a 'fail: command not found' error. By replacing this
call for a 'complain()' call, the command not found error is not longer
thrown and the right error is printed.

WIP #696

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>