src: kernel_config_manager: fix hanging .config files
When the 'kw --save | -s' fails because the commit in the 'kw/configs'
repo could not be made, the .config file and its metadata are added to
the management of kw anyways. By hard resetting the 'kw/config' repo,
the lingering files are removed and the repo stays consistent with the
error thrown to the user.

Closes #696

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>