src: kernel_config_manager: remove undesirable output
When a commit in the kw .config file manager fails, the abortion of the
save operation is made with a git reset --hard. However, the output of
this command is printed in the user stdout. With this commit, this
output is supressed as it makes no sense to the user.

WIP #696

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>