tests: remote: add test for ssh_connection_failure_message function
Although this function does not influence in the logic of the module 'kw
ssh', it displays useful information for the user in case of a failed
connection (mainly the IP, user and port attempted). As such, a test to
check if the correct information is being display in some failure cases
is important.

WIP #695

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>