tests: remote: add tests for is_ssh_connection_configured function
This important function, as it is called for every 'kw ssh' call, did
not have a test. As such, we included a unit test for it considering the
three cases of an attempted connection: 1) explicitly stating the host
with 'kw ssh -r'; 2) using a remote config file; 3) using a remote
config file, but kw could not find one.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>
Co-authored-by: Everaldo Gomes Junior <everaldogjr@gmail.com>