setup.sh: fix installed package detection
Do not use regex to query if a package is installed on archlinux. This
avoids false-positives.

As an example, the package `bc` is said to be installed if the package
`curl` is installed. The `bc` is particulary plobematic because of it
small name but could be other combinations of installed packages and kw
dependencies which could cause false positives.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: Pedro Rabello Sato <kamorst@usp.br>