src: remote: fix error message from failed ssh connection
When the user tries to connect via 'kw ssh' using a remote config file
instead of explicitly typing the IP, user and port (with 'kw ssh -r')
and fails, those three informations are not displayed to the user. This
was probably due to the change from adding this explicit declaration of
the host. The 'ssh_connection_failure_message' function expects that
those informations are already set in the 'remote_parameters' variable.
In the case of using a remote config file, these are not set and we need
to retrieve them from the file to avoid them being displayed as blank.
Also added an error message if no remote config file is found.

Closes #695

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>