src: remote: handle error of remote config file not found
In the 'is_ssh_connection_configured' function there was no error
handling when kw did not find a remote config file. With this commit,
rather than ignoring this case and trying to establish a connection
anyway we return an error code of no entry (ENOENT) and avoid extra
computation when we know that it will fail.

Co-authored-by: Everaldo Gomes Junior <everaldogjr@gmail.com>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>