src: deploy: Decompose run_kernel_install in small functions
The function run_kernel_install has all the required steps for preparing
a kernel to be deployed; however, this function is huge and hard to
reuse. As a preparation to enable support for creating kw packages, this
commit decouples run_kernel_install in multiple tiny functions and
creates the function responsible for building the kw package.

Closes  #707

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>