src: kernel_config_manager.sh: Remove git dependency
The KW subcommand kernel_config_manager used a rather complex scheme to
track .config files. Now, the management is done using the file system
with common commands (like "rm" and "cp"). This will simplify and remove
some complexity from the code.

It is added a function in the setup.sh script to deal with old kw
installations that used git to manager .config files.

Reviewed-by: Rubens Gomes Neto <rubens.gomes.neto@alumni.usp.br>
Signed-off-by: Pedro Rabello Sato <kamorst@usp.br>