src: config: add configurations listing
This commit adds a feature to kw config of listing configurations. This
listing can be of all possible config files (vm, build, deploy, mail,
notification and kworkflow) or of a subset of those. The information is
displayed in a similar fashion to the git config --list command.

To list all configurations use: kw (config | config --show | config -s)
To list a subset use: kw (config --show | config -s) <config_file>...

Closes #701

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>