src: update: add self-update mechanism to kw
Now the user can update kw by just running 'kw self-update' instead of
having to go to a kw repo, checking out to the given branch and running
the setup.sh script. By default, 'kw update' clones the official kw repo
a the temporary directory and installs it from the master branch. Also,
if the user wants to install the current unstable version of kw, the
flag '--unstable' can be used.

Closes #303

Co-authored-by: Everaldo Gomes Junior <everaldogjr@gmail.com>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>