src: lib: dialog_ui: Add basic support for dialog
We want to add support for kw to interact with lore interface. This
commit is responsible for adding a new lib folder where it saves the
dialog lib file. Notice that this is a dummy change just to enable some
basic validation and create a small foundation for the next part of this
work.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>