src: lib: upstream_patches_ui: Add new screens
For the kw upstream UI, we created a tiny mockup that shows our basic
goal with this interface which can be checked at:

 https://gitlab.freedesktop.org/siqueira/lore-prototype/-/issues/4

This commit leverages the already available dialog functions to
implement the following screens: bookmark, list patches, and register
mailing list. In the process of adding those new screens, it was
required to expand the dialog library to support more options.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>