src: deploy: Fix individual redirects to same file
Previously, there were individual printf redirects to the same file,
which conflicted with rule SC2129 of ShellCheck that says to bundle
those redirects in only one (https://www.shellcheck.net/wiki/SC2129).
The reasoning is that the file is opened and closed only once and thus
there is a performance gain, although, in this case, the gain is
irrelevant being more an issue of keeping the project coding style.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>