src: deploy: Implement deploy commands outside kernel tree
Some kw deploy commands, such as --list and --uninstall, should be able
to be executed outside a kernel tree.

The implementation revolves around these commands enabling a flag that
does not trigger the checking of an existing kernel tree in the current
directory (like the --from-package deploy command).

Important to say that the test cases associated with this implementation
only check if the flag has been set, because the checking of an existing
kernel tree is done in the deploy_main() which does not make sense to
test.

Closes #734

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>