src: kw_remote: Fix not failing when missing valid options
Before this commit, kw remote did not validate if the command had valid
option, i.e., calling kw remote (without option/parameters) or calling
kw remote <name> <remote> (without valid option) did not fail, although
it did not seems to have any side effect.

Now we validate if any of the options add, remove, rename, list or set
default remote were passed at the end of parse_remote_options(). If not,
the command fails.

Chose to validate at the end of parse_remote_options() as 1) it seems
the responsability of this function and 2) it enables to make a simple
unit test for this case.

Closes #728

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>