documentation: man: remote: Revise kw remote man page
This commit corrects some typos, grammar errors and reinforces the
project's standards for the kw remote man page. The kw mail man page was
used as a reference for these changes, as kw mail is one of the newest
features of kw and its man page is a great example of documentation.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>