src: _kw: Add ZSH completion for kw main commands
This commit changes the ZSH completion from an adaptation of the Bash
completion to a native implementation. As mentioned before, the command
'mapfile' by the Bash implementation did not adapted well and broke the
ZSH completion.

The approach was to build a native completion system for ZSH of kw. This
was done by creating a new source file '_kw' to hold the completion
functions and changing what 'setup.sh' appended to .zshrc (ZSH config
file). This commit only implements the main dispatcher, i.e., the
function which completes the main kw commands/modules and delegates the
subsequent completions to the specific function (which none are
implemented at the moment).

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>