src: kw_remote: Add support for global remote.config
Although there was a global remote.config inside kw configs directory,
there was no way to manage it through the kw remote feature.

This commit adds support to manage the global remote.config through kw
remote. The choice between which remote.config to use is based on the
existence or not of the local and global directories and files. The
local ones have precedence over the global ones and the choice is
transparent to the user. In practical terms, if we use kw remote inside
a directory that contains the .kw directory the local scope is going to
be used. Otherwise, the global scope will be used.

Closes #775

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Co-authored-by: Pedro Rabello Sato <kamorst@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>