src: kw_env: fixing the "env.current does not exist" bug.
The destroy_env() function had a bug where it looked for the env.current
file, and that file didn't exist. This bug occurred when the user created
an env and then destroyed that same env.

This commit checks that the env.current file exists to see if the user is
using the environment he wants to destroy.

Closes: #774

Signed-off-by: Aquila Macedo <aquilamacedo@riseup.net>