tests: kwio_test: Use echo instead of printf
In general, we avoid the use of echo, but there is one case in which
echo can be better, and it is in one of the subtests for kwio.
Basically, when running those tests in a bare minimal chroot, it is
possible to see this failure:

 test_alert_completion_visual_alert
 ASSERT:Variable v should exist. expected:<TESTING COMMAND> but was:<The following command set in the visual_alert_command variable could not be run:
 /bin/printf "%s
 " "$COMMAND"
 Check if the necessary packages are installed.>
 shunit2:ERROR test_alert_completion_visual_alert() returned non-zero return code.
 test_alert_completion_sound_alert
 ASSERT:Variable s should exist. expected:<TESTING COMMAND> but was:<The following command set in the sound_alert_command variable could not be run:
 /bin/printf "%s
 " "$COMMAND"
 Check if the necessary packages are installed.>
 shunit2:ERROR test_alert_completion_sound_alert() returned non-zero return code.

The reason for that came from the fact that /bin/printf does not exist
in the chroot; however, /bin/echo exists in almost every distro. This
commit replaces printf only in the test_alert_completion_visual_alert to
ensure this test passes in multiple conditions.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>