tests: maintainers_test: Add get_maintainer.pl in the samples folder
maintainers_test was written with the external dependency of
get_maintainer.pl, which makes the test fragile and adds the requirement
of downloading get_maintainer before running the tests. This commit
removes the external dependency by copying one specific version of the
get_maintainer file to the samples folder. This solution is more like a
workaround, and the maintainers_test should be rewritten in the near
future.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>