src: kw: Check if kw is installed at /
In the Debian package version of kw, a patch makes the variable
KW_LIB_DIR point to /usr/share/kw. The Debian patch exposes a problem
with how kw deals with default paths. To address this issue, this commit
adds a verification path in the kw file; if kw is installed in the
system kw will prefer to use the system-wide version.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>