src: deploy_test: Explicitly choose filesystem type
The `test_prepare_distro_for_deploy` expected the ext4 filesystem from
the host. If the host is btrfs this functions runs `btrfs property get /
ro | grep "ro=false" --silent` which fails inside a ext4 host.
Modify the test to explicitly mock the filesystem with etx4 and add a
test to mock btrfs.

Signed-off-by: Pedro Rabello Sato <kamorst@usp.br>