documentation: dependencies: Add curl and xpath dependencies
The commands curl and xpath are used by the kw upstream-patches-ui
command, but those dependencies are not checked when kw is installed.
Although they are really common commands in almost every linux
installation, if the setup doesn't have these dependencies installed,
a command not found error occurs.

This commit fix this problem by adding both dependencies to the
dependencies files of Arch, Debian and Fedora distros.

Closes: #792

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>