src: upstream_patches_ui: Add help option
The kw upstream_patches_ui doesn't have a help option to inform the user
about the command usage.

This commit adds a -h option that superficially tells what the feature
is about.

Note: This commit also adds a --help option to show the man page of the
feature, but, as there isn't a man page yet, using this option results
in nothing. Also, the implementation follows the one done in the kw mail
feature, as it is the most up-to-date one.

Closes: #793

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>