src: upstream_patches_ui: Add loading screen for delayed actions
When the user tries to access patches from the public mailing list or
download a given patch (or series) there is a possible delay between the
requisiton and the completion (i.e. the display of the new screen).
There is no visual notification to the user (for the dialog view there
is only a black screen with a cursor) which does not indicate whether
an error happened or if it is just a delay.

This commit adds this visual notification with the addition of the
create_loading_screen function that accepts a custom message to be
displayed to the user. In case of the dialog view, this message is
displayed inside an info box.

Note that this function may be expanded to display a progress bar or
gauge.

Closes: #797

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>