src: upstream_patches_ui: Add bookmark feature
This commit adds a barebones implementation of the bookmark feature.
This feature refers to the ability to choose a certain patch series from
a public mailing list and save it to later consultation. After
bookmarking a patch series, it will become available in the 'Bookmarked
Patches' screen. At the moment, the only action available is to
'Unbookmark' a patch series, but it should be kind of simple to plug in
new actions.

From the implementation point of view, kw needs to maintain a local
database of all bookmarked patch series and this is done by creating and
managing a file inside the kw data directory (much like kw
kernel-config-manager). Although efficient and simple to manage, if the
feature needs scaling (change/add entities/fields), a real DBMS should
be considered. Also worth mentioning that, under the hood, bookmarking a
patch series downloads it first in the same manner of the 'Download'
action, i.e., saves the .mbox files inside a directory defined by the
user in lore.config. Unbookmarking patches series also removes them
first from local storage.

Closes: #796

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>