src: upstream_patches_ui: Add message box screen
When kw tries to call list_patches, but there are no patches to be
displayed (i.e., the _target_array_list is empty), dialog just hangs.
Furthermore, there isn't any information displayed to the user regarding
the occurrence.

In this context, this commit adds a message box screen that can, not
only be used to fix this case above, but also be used for any exception
treatment or notification. For dialog, when this screen occurs, a custom
title and message and an OK button to return to the flow of execution
are displayed.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>