src: lib: dialog_ui: Add persistence to checklist options
When the user bookmarked a patch (series) the next screen didn't alter
the status of the chosen options. In this case, we want the 'Bookmark'
and 'Download' options already checked, to notify the user that the
bookmark worked (and also for futher visits to the patch).

In this context, this commits implements this behaviour by adding
another parameter to the create_simple_checklist function named
check_statuses that is an array reference with all the statuses of the
checks (if they are on/off). As such, this behaviour is not limited to
the 'Bookmark' and 'Download' actions on a patch (series).

Notes:
- As we plan to revise the 'Download' action, when we do it, we need to
  decouple the 'Bookmark' check status from the 'Download' one (right
  now, 'Download' only checks when 'Bookmark' is checked);
- We may want to reduce the number of parameters in the
  create_simple_checklist function (and others in dialog_ui.sh) as it
  reached 10 arguments with this commit.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>