src: upstream_patches_ui: Fix Dashboard screen message box
The Dashboard screen message box isn't related to the screen. This
commit fixes this by making the message box empty (no message), as the
title 'Dashboard' is already indicative and this is a generic screen
that assumes the menu options are self-descriptive. Also declared the
variable used 'message_box' for safety keeping.

Closes: #805

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>