src: kwlib: Add function for safety checking paths
It is a common task to remove files and directories. To avoid dangerous
removals, we need to check if the path is safe to remove. This commit
adds a predicate function named is_safe_path_to_remove that checks a
path safety.

An alternate approach would be to also encapsulate the removal, but
the chosen one gives the caller more flexibility.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>