src: upstream_patches_ui: Add apply patch action
It would be good if the user could choose a patch (series) and apply it
to a given kernel tree source and branch via upstream-patches-ui. This
commit add a barebones implementation of this feature. To apply the
patch, kw uses the values of 'kernel_tree' and 'target_branch' of
lore.config.

Note that this is a draft of the feature as there are no documentation,
error handling or safe checking.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>